requirements.txt
asgiref==3.4.1
Django==3.2.9
pytz==2021.3
sqlparse==0.4.2
drf-yasg==1.20.0
markdown
django-filter
django-rest-knox
